Revert "AtomicContainer"
This reverts commit ad342d17d75f9b6d02f37b50ff0a4e53683eecb9.
